chore(release): Add all distribution files for jsDelivr updater
jsDelivr has an auto-update feature based on github, so it will be more conviennent to keep all distribution files in a 'dist' folder.

As jsDelivr will update on each new tag, distribution files (minified JS/CSS) have been moved to a new grunt task named 'dist', so it won't cause conflicts when in development. Non minified files are still build in default task and placed in 'assets' directoy for bower #master to work properly. 'dist' task and population of dist folder must be performed only for a new release.