refactor(css): Apply row and tasks model classes CSS to containers.
Classes defined in row and tasks model are applied to all containers of objects instead of background only. This allows more possible customization, because a task/row is displayed using multiple divs (background, foreground, content ...).

This will break existing styling based on those custom model CSS classes, but can be fixed easily by updating selectors in the stylesheet.

For example, if a class is defined on task to change background color

custom-task-look {
    background: red;
}

Needs to be replaced with

custom-task-look gantt-task-background {
    background: red;
}

Read docs for more information (data page).