perf(tasks): Increase performance by not adding new tasks to the visible/filtered task lists.
Adding many new tasks is slow because new tasks are added to the visible/filtered task lists. Because of that Angular will initiate the scope+elements and draw them until the filter kicks in by e.g. scrolling or some other actions. Usually this is not visible to the user but affects performance. The proposed change will not automatically add a new task to those lists. It uses the existing updateVisibleTasks functionality instead.