perf(scroll): Remove hidden flag filter
The hidden flag filter uses the default Angular filter which is not very fast as an object comparison is done to determine if the hidden flag is true/false. Additionally the hidden flag requires to go trough the array of columns/headers/rows/tasks three times (1st to set hidden=true, 2nd to set hidden = false if match, 3rd to apply the filter inside the template). By having a own property visibleColumns/Header/Rows/Tasks no filter is necessary and the scrolling performance is double as fast.