refac(gantt): Removed element-height-listener directive and one element-width-listener
Remove element-listeners except for one because it is needed by one last `bodyRowsWidth` watcher. Unfortunately when removing `bodyRowsWidth` out of this watcher there will be a display issue when hide/show the gantt labels. The scroll code is refactored too to use a isolate scope and access the body width directly using the element width.