refac(gantt): Use scope value for side section and gantt width to prevent display issues
When mixing direct DOM properties with scope variables the DOM values may still be old as the apply cycle is not finished. Or the DOM properties are already new as they got set directly while others DOM properties get set with scope variables and therefore the gantt does not behave correctly. This can lead to undesired results as described in #243.
With the refactoring the gantt width is based on the ganttElementWidth and sideWidth scope variable. All other main sizes can be calculated based on this two values. Through this the autoFitWidth is working much better and the bodyRowWidth/bodyRowHeight element listener could be removed.
 As a scope value is needed for the side section the section is 150px wide by default. The width can be either changed by resizing the side section, with api methods or by calling api.label.fitSideWidth to automatically adjust the side width to the labels width. As the side section belongs to the gantt a new gantt options has been introduced: 'show-side' to hide/show the side section. The side section is not shown when no side section plugin is enabled or loaded. This behaviour would also allow to add multiple side section plugins.

 Fixes #243